Reorganize code in styling module
Improved naming of modules in gaphor.core.styling.

Get rid of the parser module, since it's name is confusing with the
tinycss2 parser module.
