+++ gaphor/core/styling/selectors.py
@@ -1,221 +1,412 @@
-"""Compiler for CSS selectors.
+"""Parser for CSS selectors, based on cssselect2.
 
-Based on cssselect2.compiler, written by Simon Sapin and Guillaume
-Ayoub.
+Changes compared to cssselect2:
+
+* The parser has been adapted to support statements like `class[subject.ownedAttribute=foo]`.
+* :not is treated as a normal pseudo-class. It's a `FunctionalPseudoClassSelector`.
+
+Original module: cssselect2.parser
+:copyright: (c) 2012 by Simon Sapin, 2017 by Guillaume Ayoub.
+:license: BSD, see https://github.com/Kozea/cssselect2/blob/master/LICENSE for more details.
 """
-
-import re
-from functools import singledispatch
-from typing import Callable, Dict, Iterator, Literal, Tuple, Union
 
 import tinycss2
 
-from gaphor.core.styling import parser
-from gaphor.core.styling.declarations import parse_declarations
 
-# http://dev.w3.org/csswg/selectors/#whitespace
-split_whitespace = re.compile("[^ \t\r\n\f]+").findall
-
-
-Rule = Union[
-    Tuple[Tuple[Callable[[object], bool], Tuple[int, int, int]], Dict[str, object]],
-    Tuple[Literal["error"], Union[tinycss2.ast.ParseError, parser.SelectorError]],
-]
-
-
-def compile_style_sheet(*css: str) -> Iterator[Rule]:
-    for sheet in css:
-        if sheet:
-            rules = tinycss2.parse_stylesheet(
-                sheet, skip_comments=True, skip_whitespace=True
-            )
-            yield from compile_rules(rules)
+def selectors(input, namespaces=None):
+    """Parse tokens or string into selectors."""
+    if isinstance(input, str):
+        input = tinycss2.tokenizer.parse_component_value_list(input, skip_comments=True)
+    tokens = TokenStream(input)
+    namespaces = namespaces or {}
+    yield parse_selector(tokens, namespaces)
+    tokens.skip_whitespace_and_comment()
+    while 1:
+        next = tokens.next()
+        if next is None:
+            return
+        elif next == ",":
+            yield parse_selector(tokens, namespaces)
+        else:
+            raise SelectorError(next, f"unexpected {next.type} token.")
 
 
-def compile_rules(rules):
-    for rule in rules:
-        if rule.type == "error":
-            yield "error", rule
-            continue
+def media_query_selector(input):
+    tokens = TokenStream(input)
+    query = []
+    while 1:
+        tokens.skip_whitespace_and_comment()
+        next = tokens.next()
+        if next is None:
+            break
+        elif next.type in ("ident", "literal", "string"):
+            query.append(next.value)
+        elif next.type == "() block":
+            # lazy: only parse inside the parenthesis
+            # we do not support combinations at this time
+            tokens = TokenStream(next.content)
+        else:
+            raise SelectorError(next, f"unexpected {next.type} token.")
 
-        if rule.type == "at-rule" and rule.content:
-            at_rules = tinycss2.parser.parse_rule_list(
-                rule.content, skip_comments=True, skip_whitespace=True
-            )
-            media_selector = parser.parse_media_query(rule.prelude)
-            if not media_selector:
-                continue
-            media_query = compile_node(media_selector)
-            yield from (
-                ((_combine(media_query, selspec[0]), selspec[1]), declaration)
-                for selspec, declaration in compile_rules(at_rules)
-                if selspec != "error"
-            )
-
-        if rule.type != "qualified-rule":
-            continue
-
-        try:
-            selectors = compile_selector_list(rule.prelude)
-        except parser.SelectorError as e:
-            yield "error", e
-            continue
-
-        declaration = {
-            prop: value
-            for prop, value in parse_declarations(rule.content)
-            if prop != "error" and value is not None
-        }
-
-        yield from ((selector, declaration) for selector in selectors)
+    return MediaSelector(query)
 
 
-def _combine(a, b):
-    return lambda el: a(el) and b(el)
+def parse_selector(tokens, namespaces):
+    result = parse_compound_selector(tokens, namespaces)
+    while 1:
+        has_whitespace = tokens.skip_whitespace()
+        while tokens.skip_comment():
+            has_whitespace = tokens.skip_whitespace() or has_whitespace
+        peek = tokens.peek()
+        if peek in (">", "+", "~"):
+            combinator = peek.value
+            tokens.next()
+        elif peek is None or peek == "," or not has_whitespace:
+            return result
+        else:
+            combinator = " "
+        compound = parse_compound_selector(tokens, namespaces)
+        result = CombinedSelector(result, combinator, compound)
 
 
-def compile_selector_list(input):
-    """Compile a (comma-separated) list of selectors.
+def parse_compound_selector(tokens, namespaces):
+    type_selectors = parse_type_selector(tokens, namespaces)
+    simple_selectors = type_selectors if type_selectors is not None else []
+    while 1:
+        simple_selector = parse_simple_selector(tokens, namespaces)
+        if simple_selector is None:
+            break
+        simple_selectors.append(simple_selector)
 
-    Based on cssselect2.compiler.compile_selector_list().
+    if simple_selectors or type_selectors is not None:
+        return CompoundSelector(simple_selectors)
+    peek = tokens.peek()
+    raise SelectorError(
+        peek, f'expected a compound selector, got {peek.type if peek else "EOF"}'
+    )
 
-    Returns a list of compiled selectors.
+
+def parse_type_selector(tokens, namespaces):
+    tokens.skip_whitespace()
+    qualified_name = parse_qualified_name(tokens, namespaces)
+    if qualified_name is None:
+        return None
+
+    simple_selectors = []
+    namespace, local_name = qualified_name
+    if local_name is not None:
+        simple_selectors.append(LocalNameSelector(local_name))
+    if namespace is not None:
+        simple_selectors.append(NamespaceSelector(namespace))  # type: ignore[arg-type]
+    return simple_selectors
+
+
+def parse_simple_selector(tokens, namespaces):
+    peek = tokens.peek()
+    if peek is None:
+        return None
+    if peek.type == "hash" and peek.is_identifier:
+        tokens.next()
+        return IDSelector(peek.value)
+    elif peek == ".":
+        tokens.next()
+        next = tokens.next()
+        if next is None or next.type != "ident":
+            raise SelectorError(next, f"Expected a class name, got {next}")
+        return ClassSelector(next.value)
+    elif peek.type == "[] block":
+        tokens.next()
+        return parse_attribute_selector(TokenStream(peek.content), namespaces)
+    elif peek == ":":
+        tokens.next()
+        next = tokens.next()
+        if next == ":":
+            next = tokens.next()
+            if next is None or next.type != "ident":
+                raise SelectorError(next, f"Expected a pseudo-element name, got {next}")
+            return PseudoElementSelector(next.lower_value)
+        elif next is not None and next.type == "ident":
+            return PseudoClassSelector(next.lower_value)
+        elif next is not None and next.type == "function":
+            return FunctionalPseudoClassSelector(next.lower_name, next.arguments)
+        else:
+            raise SelectorError(next, f"unexpected {next} token.")
+    else:
+        return None
+
+
+def parse_attribute_selector(tokens, namespaces):
+    tokens.skip_whitespace()
+    qualified_name = parse_qualified_name(tokens, namespaces, is_attribute=True)
+    if qualified_name is None:
+        next = tokens.next()
+        raise SelectorError(next, f"expected attribute name, got {next}")
+    namespace, local_name = qualified_name
+
+    # Allow syntax like "subject.ownedAttribute"
+    if local_name:
+        name, lower_name = local_name
+        while 1:
+            peek = tokens.peek()
+            if peek == ".":
+                next = tokens.next()
+                name += next.value
+                lower_name += next.value
+            elif peek and peek.type == "ident":
+                next = tokens.next()
+                name += next.value
+                lower_name += next.lower_value
+            else:
+                break
+
+        local_name = name, lower_name
+
+    tokens.skip_whitespace()
+    peek = tokens.peek()
+    if peek is None:
+        operator = None
+        value = None
+    elif peek in ("=", "~=", "|=", "^=", "$=", "*="):
+        operator = peek.value
+        tokens.next()
+        tokens.skip_whitespace()
+        next = tokens.next()
+        if next is None or next.type not in ("ident", "string"):
+            next_type = "None" if next is None else next.type
+            raise SelectorError(next, f"expected attribute value, got {next_type}")
+        value = next.value
+    else:
+        raise SelectorError(peek, f"expected attribute selector operator, got {peek}")
+
+    tokens.skip_whitespace()
+    next = tokens.next()
+    if next is not None:
+        raise SelectorError(next, f"expected ], got {next.type}")
+    return AttributeSelector(namespace, local_name, operator, value)
+
+
+def parse_qualified_name(tokens, namespaces, is_attribute=False):
+    """Return ``(namespace, local)`` for given tokens.
+
+    Can also return ``None`` for a wildcard.
+
+    The empty string for ``namespace`` means "no namespace".
     """
-    return [
-        (compile_node(selector), selector.specificity)
-        for selector in parser.parse(input)
-    ]
+    peek = tokens.peek()
+    if peek is None:
+        return None
+    if peek.type == "ident":
+        first_ident = tokens.next()
+        peek = tokens.peek()
+        if peek != "|":
+            namespace = "" if is_attribute else namespaces.get(None, None)
+            return namespace, (first_ident.value, first_ident.lower_value)
+        tokens.next()
+        namespace = namespaces.get(first_ident.value)
+        if namespace is None:
+            raise SelectorError(
+                first_ident, f"undefined namespace prefix: {first_ident.value}"
+            )
+
+    elif peek == "*":
+        next = tokens.next()
+        peek = tokens.peek()
+        if peek != "|":
+            if is_attribute:
+                raise SelectorError(next, f"Expected local name, got {next.type}")
+            return namespaces.get(None, None), None
+        tokens.next()
+        namespace = None
+    elif peek == "|":
+        tokens.next()
+        namespace = ""
+    else:
+        return None
+
+    # If we get here, we just consumed '|' and set ``namespace``
+    next = tokens.next()
+    if next.type == "ident":
+        return namespace, (next.value, next.lower_value)
+    elif next == "*" and not is_attribute:
+        return namespace, None
+    else:
+        raise SelectorError(next, f"Expected local name, got {next.type}")
 
 
-@singledispatch
-def compile_node(selector):
-    """Dynamic dispatch selector nodes.
+class SelectorError(ValueError):
+    """A specialized ``ValueError`` for invalid selectors."""
 
-    Default behavior is to deny (no match).
+
+class TokenStream:
+    def __init__(self, tokens):
+        self.tokens = iter(tokens)
+        self.peeked = []  # In reversed order
+
+    def next(self):
+        return self.peeked.pop() if self.peeked else next(self.tokens, None)
+
+    def peek(self):
+        if not self.peeked:
+            self.peeked.append(next(self.tokens, None))
+        return self.peeked[-1]
+
+    def skip(self, skip_types):
+        found = False
+        while 1:
+            peek = self.peek()
+            if peek is None or peek.type not in skip_types:
+                break
+            self.next()
+            found = True
+        return found
+
+    def skip_whitespace(self):
+        return self.skip(["whitespace"])
+
+    def skip_comment(self):
+        return self.skip(["comment"])
+
+    def skip_whitespace_and_comment(self):
+        return self.skip(["comment", "whitespace"])
+
+
+class MediaSelector:
+    def __init__(self, query):
+        self.query = query
+
+    @property
+    def specificity(self):
+        return 0, 0, 0
+
+    def __repr__(self):
+        return "".join(self.query)
+
+
+class CombinedSelector:
+    def __init__(self, left, combinator, right):
+        #: Combined or compound selector
+        self.left = left
+        # One of `` `` (a single space), ``>``, ``+`` or ``~``.
+        self.combinator = combinator
+        #: compound selector
+        self.right = right
+
+    @property
+    def specificity(self):
+        a1, b1, c1 = self.left.specificity
+        a2, b2, c2 = self.right.specificity
+        return a1 + a2, b1 + b2, c1 + c2
+
+    def __repr__(self):
+        return "{!r}{}{!r}".format(self.left, self.combinator, self.right)
+
+
+class CompoundSelector:
+    """Aka.
+
+    sequence of simple selectors, in Level 3.
     """
-    raise parser.SelectorError("Unknown selector", selector)
+
+    def __init__(self, simple_selectors):
+        self.simple_selectors = simple_selectors
+
+    @property
+    def specificity(self):
+        if self.simple_selectors:
+            return tuple(
+                map(sum, zip(*(sel.specificity for sel in self.simple_selectors)))
+            )
+        else:
+            return 0, 0, 0
+
+    def __repr__(self):
+        return "".join(map(repr, self.simple_selectors))
 
 
-@compile_node.register
-def compile_media_selector(selector: parser.MediaSelector):
-    query = selector.query
-    if len(query) == 1:
-        mode = query[0].lower()
-    elif (
-        len(query) == 3
-        and query[0].lower() == "prefers-color-scheme"
-        and query[1] == "="
-    ):
-        mode = query[2].lower()
-    else:
-        mode = None
+class LocalNameSelector:
+    specificity = 0, 0, 1
 
-    if mode in ("dark", "dark-mode"):
-        return lambda el: el.dark_mode is True
-    elif mode in ("light", "light-mode"):
-        return lambda el: el.dark_mode is False
-    return lambda el: False
+    def __init__(self, local_name):
+        self.local_name, self.lower_local_name = local_name
+
+    def __repr__(self):
+        return self.local_name
 
 
-@compile_node.register
-def compile_compound_selector(selector: parser.CompoundSelector):
-    sub_expressions = [compile_node(sel) for sel in selector.simple_selectors]
-    return lambda el: all(expr(el) for expr in sub_expressions)
+class NamespaceSelector:
+    specificity = 0, 0, 0
+
+    def __init__(self, namespace):
+        #: The namespace URL as a string,
+        #: or the empty string for elements not in any namespace.
+        self.namespace = namespace
+
+    def __repr__(self):
+        return "|" if self.namespace == "" else "{%s}|" % self.namespace
 
 
-@compile_node.register
-def compile_name_selector(selector: parser.LocalNameSelector):
-    return lambda el: el.name() == selector.lower_local_name
+class IDSelector:
+    specificity = 1, 0, 0
+
+    def __init__(self, ident):
+        self.ident = ident
+
+    def __repr__(self):
+        return f"#{self.ident}"
 
 
-def ancestors(el):
-    if p := el.parent():
-        yield p
-        yield from ancestors(p)
+class ClassSelector:
+    specificity = 0, 1, 0
+
+    def __init__(self, class_name):
+        self.class_name = class_name
+
+    def __repr__(self):
+        return f".{self.class_name}"
 
 
-def descendants(el):
-    for c in el.children():
-        yield c
-        yield from descendants(c)
+class AttributeSelector:
+    specificity = 0, 1, 0
+
+    def __init__(self, namespace, name, operator, value):
+        self.namespace = namespace
+        self.name, self.lower_name = name
+        #: A string like ``=`` or ``~=``, or None for ``[attr]`` selectors
+        self.operator = operator
+        #: A string, or None for ``[attr]`` selectors
+        self.value = value
+
+    def __repr__(self):
+        namespace = "*|" if self.namespace is None else "{%s}" % self.namespace
+        return "[{}{}{}{!r}]".format(namespace, self.name, self.operator, self.value)
 
 
-@compile_node.register
-def compile_combined_selector(selector: parser.CombinedSelector):
-    left_inside = compile_node(selector.left)
-    if selector.combinator == " ":
+class PseudoClassSelector:
+    specificity = 0, 1, 0
 
-        def left(el):
-            return any(left_inside(e) for e in ancestors(el))
+    def __init__(self, name):
+        self.name = name
 
-    elif selector.combinator == ">":
-
-        def left(el):
-            p = el.parent()
-            return p is not None and left_inside(p)
-
-    else:
-        raise parser.SelectorError("Unknown combinator", selector.combinator)
-
-    right = compile_node(selector.right)
-    return lambda el: right(el) and left(el)
+    def __repr__(self):
+        return f":{self.name}"
 
 
-@compile_node.register
-def compile_attribute_selector(selector: parser.AttributeSelector):
-    name = selector.lower_name
-    operator = selector.operator
-    value = selector.value and selector.value.lower()
+class PseudoElementSelector:
+    specificity = 0, 0, 1
 
-    if operator is None:
-        return lambda el: bool(el.attribute(name))
-    elif operator == "=":
-        return lambda el: el.attribute(name) == value
-    elif operator == "~=":
-        return lambda el: value in split_whitespace(el.attribute(name))
-    elif operator == "^=":
-        return lambda el: value and el.attribute(name).startswith(value)
-    elif operator == "$=":
-        return lambda el: value and el.attribute(name).endswith(value)
-    elif operator == "*=":
-        return lambda el: value and value in el.attribute(name)
-    elif operator == "|=":
+    def __init__(self, name):
+        self.name = name
 
-        def pipe_equal_matcher(el):
-            v = el.attribute(name)
-            return v == value or v and v.startswith(f"{value}-")
-
-        return pipe_equal_matcher
-    else:
-        raise parser.SelectorError("Unknown attribute operator", operator)
+    def __repr__(self):
+        return f"::{self.name}"
 
 
-@compile_node.register
-def compile_pseudo_class_selector(selector: parser.PseudoClassSelector):
-    name = selector.name
-    if name == "empty":
-        return lambda el: not next(el.children(), 0)
-    elif name in ("root", "hover", "focus", "active", "drop", "disabled"):
-        return lambda el: name in el.state()
-    else:
-        raise parser.SelectorError("Unknown pseudo-class", name)
+class FunctionalPseudoClassSelector:
+    specificity = 0, 1, 0
 
+    def __init__(self, name, arguments):
+        self.name = name
+        self.arguments = arguments
 
-@compile_node.register
-def compile_functional_pseudo_class_selector(
-    selector: parser.FunctionalPseudoClassSelector,
-):
-    name = selector.name
-    if name not in ("has", "is", "not"):
-        raise parser.SelectorError("Unknown pseudo-class", name)
-
-    sub_selectors = compile_selector_list(selector.arguments)
-    selector.specificity = max(spec for _, spec in sub_selectors)
-    if name == "has":
-        return lambda el: any(
-            any(sel(c) for sel, _ in sub_selectors) for c in descendants(el)
-        )
-    elif name == "is":
-        return lambda el: any(sel(el) for sel, _ in sub_selectors)
-    elif name == "not":
-        return lambda el: not any(sel(el) for sel, _ in sub_selectors)
+    def __repr__(self):
+        return ":{}{!r}".format(self.name, tuple(self.arguments))
