+++ gaphor/core/styling/tests/test_declarations.py
